Allow SHA-256 and SHA-512 as salt for crypt
Currently the best that this PAM module can offer for password
encryption is MD5. In cybersecurity, MD5 has long been considered too
unsafe!
This patch therefore adds the option, if the system supports it,
to encrypting passwords in the SQLite with SHA256 or SHA512,
both are currently classified as secure by the cyber security
community.