add function memzero_explicit
In the file pam_get_passwd.c, the dynamic memory in which passwords are
stored is overwritten with 0 before the memory is released but
modern compilers and linkers recognize that the memset is called
directly before a free and try to opitmize by removing this call.

In order not to remove, the memset is protected in the new function
memzero_explicit with a memory barrier, so that the compiler is forced
to leave this in code in the binary.