allow ciphers sha-256 and sha-512 in pam_sqlite  (#8)
* Always fill result in crypt_make_salt with random

this makes it easier to implement later other chipers as well.like
SHA-256 or SHA-512. In addition, a constant time behavior ensures that
timing attacks are made more difficult, otherwise a timing attack
against crypt_make_salt would be possible to draw conclusions about
the selected chiper.

* According to the crypt manpage,

a salt can be up to 16 characters long. This should also be used,
because if the salt is too small, it may be worthwhile for an attacker
to calculate a hash table despite the salts.

* Allow SHA-256 and SHA-512 as salt for crypt

Currently the best that this PAM module can offer for password
encryption is MD5. In cybersecurity, MD5 has long been considered too
unsafe!
This patch therefore adds the option, if the system supports it,
to encrypting passwords in the SQLite with SHA256 or SHA512,
both are currently classified as secure by the cyber security
community.

* Autogenerated  Recreate the ./configure file

The configure file is generated automatically with autoconf, an update
is necessary after the chipers for SHA-256 and SHA-512 have been
introduced. This is the new autogenerated version from the configure
file without any manual adjustment.

* add function memzero_explicit

In the file pam_get_passwd.c, the dynamic memory in which passwords are
stored is overwritten with 0 before the memory is released but
modern compilers and linkers recognize that the memset is called
directly before a free and try to opitmize by removing this call.

In order not to remove, the memset is protected in the new function
memzero_explicit with a memory barrier, so that the compiler is forced
to leave this in code in the binary.

* Overwrite newpass_crypt with 0 to be safe.

If the user of the module chooses PW_CLEAR as mode (default value), the
password of the user is copied into the string newpass_crypt in the
function encrypt_password. In order to prevent this password from being
leaked in the memory, it should be overwritten with 0 after use.

Co-authored-by: Oliver Schneidewind <schneidewind@printf-debugging.de>